requirements.txt
aiogram>=3.4.1,<4.0.0
aiohttp>=3.9.5,<4.0.0
python-dotenv>=1.0
